+++ axual-webclient/topic_configs.go
@@ -65,3 +65,38 @@ func (c *Client) DeleteTopicConfig(id string) error {
 	}
 	return nil
 }
+
+func (c *Client) GetTopicConfigPermissions(topicConfigID string, permType string) ([]PermissionResponse, error) {
+	var perms []PermissionResponse
+	err := c.RequestAndMap("GET", fmt.Sprintf("%s/stream_configs/%s/permissions?type=%s", c.ApiURL, topicConfigID, permType), nil, nil, &perms)
+	if err != nil {
+		return nil, err
+	}
+	return perms, nil
+}
+
+func (c *Client) DeleteTopicConfigPermissions(topicConfigID string, request PermissionRequest) error {
+	marshal, err := json.Marshal(request)
+	if err != nil {
+		return err
+	}
+	headers := map[string]string{"Content-Type": "application/json"}
+	err = c.RequestAndMap("DELETE", fmt.Sprintf("%s/stream_configs/%s/permissions?type=browse", c.ApiURL, topicConfigID), strings.NewReader(string(marshal)), headers, nil)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func (c *Client) AddTopicConfigPermissions(topicConfigID string, request PermissionRequest) error {
+	marshal, err := json.Marshal(request)
+	if err != nil {
+		return err
+	}
+	headers := map[string]string{"Content-Type": "application/json"}
+	err = c.RequestAndMap("POST", fmt.Sprintf("%s/stream_configs/%s/permissions", c.ApiURL, topicConfigID), strings.NewReader(string(marshal)), headers, nil)
+	if err != nil {
+		return err
+	}
+	return nil
+}